ToolInfo w/ JSON Schema now integrated w/ testing passing
