+++ src/inspect_ai/model/_call_tools.py
@@ -1,5 +1,4 @@
 import asyncio
-import inspect
 from dataclasses import dataclass
 from typing import (
     Any,
@@ -7,15 +6,12 @@ from typing import (
     cast,
 )
 
-from docstring_parser import Docstring, DocstringParam
-
-from inspect_ai._util.docstring import parse_docstring
 from inspect_ai._util.error import exception_message
-from inspect_ai._util.json import python_type_to_json_type
 from inspect_ai._util.registry import registry_info
-from inspect_ai.tool import Tool, ToolCall, ToolError, ToolInfo, ToolParam
+from inspect_ai.tool import Tool, ToolCall, ToolError, ToolInfo
 from inspect_ai.tool._tool import TOOL_PROMPT, ToolParsingError
 from inspect_ai.tool._tool_call import ToolCallError
+from inspect_ai.tool._tool_info import ToolParams, parse_tool_info
 
 from ._chat_message import ChatMessageAssistant, ChatMessageTool
 
@@ -85,7 +81,7 @@ class ToolDef:
     description: str
     """Tool description."""
 
-    params: list[ToolParam]
+    parameters: ToolParams
     """Tool parameters"""
 
     prompt: str | None
@@ -120,7 +116,9 @@ def tools_info(tools: list[Tool] | list[ToolInfo]) -> list[ToolInfo]:
             tdefs = tool_defs(cast(list[Tool], tools))
             return [
                 ToolInfo(
-                    name=tool.name, description=tool.description, params=tool.params
+                    name=tool.name,
+                    description=tool.description,
+                    parameters=tool.parameters,
                 )
                 for tool in tdefs
             ]
@@ -133,43 +131,21 @@ def tool_defs(tools: list[Tool]) -> list[ToolDef]:
 
 
 def tool_def(tool: Tool) -> ToolDef:
-    tool_info = registry_info(tool)
-    name = tool_info.name.split("/")[-1]
-    docstring = tool_docstring(tool)
+    name, prompt = tool_name_and_prompt(tool)
+    tool_info = parse_tool_info(tool)
 
     # build params
-    params = [tool_param(param) for param in docstring.params]
     return ToolDef(
         name=name,
-        description=str(docstring.short_description),
-        prompt=tool_info.metadata.get(TOOL_PROMPT, None),
-        params=params,
+        description=tool_info.description,
+        prompt=prompt,
+        parameters=tool_info.parameters,
         tool=tool,
     )
 
 
-def tool_param(param: DocstringParam) -> ToolParam:
-    return ToolParam(
-        name=param.arg_name,
-        type=python_type_to_json_type(param.type_name),
-        description=str(param.description),
-        optional=param.is_optional is True,
-    )
-
-
-def tool_docstring(tool: Tool) -> Docstring:
-    docstring = parse_docstring(inspect.getdoc(tool))
-    # We need tool and parameter descriptions to pass to the agent
-    assert (
-        docstring.short_description is not None
-    ), "Tool must have a short description in the docstring"
-    params = list(inspect.signature(tool).parameters.keys())
-    if len(params) > 0:
-        for param in params:
-            assert param in [
-                docstring_param.arg_name for docstring_param in docstring.params
-            ], f"Parameter {param} must be documented in the docstring"
-        assert [
-            docstring_param.description != "" for docstring_param in docstring.params
-        ], "All tool parameters must have a description"
-    return docstring
+def tool_name_and_prompt(tool: Tool) -> tuple[str, str | None]:
+    tool_registry_info = registry_info(tool)
+    name = tool_registry_info.name.split("/")[-1]
+    prompt = tool_registry_info.metadata.get(TOOL_PROMPT, None)
+    return name, prompt
