+++ src/inspect_ai/model/_providers/util.py
@@ -66,13 +66,18 @@ def parse_tool_call(
     # and then create a dict that maps it to the first function argument
     else:
         tool_info = next(
-            (tool for tool in tools if tool.name == function and len(tool.params) > 0),
+            (
+                tool
+                for tool in tools
+                if tool.name == function and len(tool.parameters.properties) > 0
+            ),
             None,
         )
         if tool_info:
             try:
                 value = yaml.safe_load(arguments)
-                arguments_dict[tool_info.params[0].name] = value
+                param_names = list(tool_info.parameters.properties.keys())
+                arguments_dict[param_names[0]] = value
             except yaml.parser.ParserError as ex:
                 report_parse_error(ex)
